removed motor encoder
Using EXTI without filtering was  too noisy. And using HW timer filtering is not portable as most low end MCUs do not support enough numbers of it or do not support it at all.

Thus, using the xHOS_RotaryEncoder is the best option as it does signal filtering and is portable.